requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Streamlit Frontend
streamlit==1.28.2
streamlit-option-menu==0.3.6

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pdfplumber==0.10.0

# NLP and ML
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# LangChain Stack
langchain==0.0.348
langchain-openai==0.0.2
langchain-community==0.0.12
langsmith==0.0.69

# Vector Databases
chromadb==0.4.18
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# OpenAI and HuggingFace
openai==1.3.7
huggingface_hub==0.19.4

# Text Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
textdistance==4.6.0
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pillow==10.1.0
plotly==5.17.0
